feat: Add SignIn component and story
The code changes add a new component called SignIn along with its corresponding story file. The SignIn component is part of the ReactComponentLibrary and is used for user sign-in functionality. The commit also includes a test case for the SignIn component.

Note: Please remove the console.log statement before committing the code.